DMT-53 Add bound-check for labels when outside of chart
Add check for responsiveness when labels are outside the chart and refactor the code by extracting the logic for labels in a separate class
